// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/dashboard.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Dashboard { 
    /**
     * Unique identifier of a dashboard.
     */
    id?: string;
    name?: string;
    definition?: string;
}